feat(080): Breaks out AbilityImprovements sub-component.
